refactor: 资源池校验优化
